Fix indentation issues
Indenting with tabs and spaces tends to end poorly - tabs are
configurable in thier width, the flow of the file was hard to read in a
cross platform way.
Artemis seems to have standadised on spaces as indentation for vesselData
files, it will be easier to keep it in sync if we copy that style.
Debatably we should just reformat the whole thing using something like
xmllint, that would get rid of the inconsistancies with the stock file,
whever thats a good or bad thing is debatable though